:art: 더이상 사용하지 않는 마크다운 마이그레이션을 지우고 toJSON을 대체
